Telemetry: summarize via cron script
